requirements.txt
flask
flask-bootstrap
flask-Login
flask-wtf
python-jsonstore
